Uebungen/24-gRPC-Introduction/Copy/src/python/requirements.txt
# tested with python 3.10.0
faker
grpcio
grpcio-tools
pandas
